+++ dtslint/ts3.5/Record.ts
@@ -79,12 +79,8 @@ _.toArray(r1) // $ExpectType ["a" | "b", number][]
 // insertAt
 //
 
-_.insertAt('b', 0)({ a: 1 }) // $ExpectType Record<"a" | "b", number>
-_.insertAt('b', 0)(l1) // $ExpectType Record<"a" | "b", number>
 _.insertAt('b', 0)(d1) // $ExpectType Record<string, number>
-_.insertAt('b', 0)(r1) // $ExpectType Record<"a" | "b", number>
 _.insertAt(stringKey, 0)(r1) // $ExpectType Record<string, number>
-_.insertAt('c', 0)(r1) // $ExpectType Record<"a" | "b" | "c", number>
 
 //
 // deleteAt
@@ -130,7 +126,7 @@ _.foldMapWithIndex(monoidString)((k: 'a' | 'b', _n) => k)(r1) // $ExpectType str
 _.reduceRightWithIndex('', (k: string, _n, _b) => k)(d1) // $ExpectType string
 _.reduceRightWithIndex('', (k: 'a' | 'b', _n, _b) => k)(r1) // $ExpectType string
 
-_.singleton('a', 1) // $ExpectType Record<"a", number>
+_.singleton('a', 1) // $ExpectType Record<string, number>
 
 _.traverseWithIndex(O.option)((_k, n: number) => O.some(n))(d1) // $ExpectType Option<Record<string, number>>
 _.traverseWithIndex(O.option)((_k: 'a' | 'b', n: number) => O.some(n))(r1) // $ExpectType Option<Record<"a" | "b", number>>